+++ test/dummy/config/routes.rb
@@ -5,6 +5,7 @@ Rails.application.routes.draw do
   # Keep legacy links working by redirecting the base path to the app home.
   get "/recording_studio", to: redirect("/"), as: nil
   mount RecordingStudio::Engine, at: "/recording_studio"
+  mount RecordingStudioAttachable::Engine, at: "/recording_studio_attachable"
   mount RecordingStudioRootSwitchable::Engine, at: "/recording_studio_root_switchable"
   mount RecordingStudioPublications::Engine, at: "/recording_studio_publications"
   recording_studio_admin_for :admin, at: "/admin", root_section: :publications